channels: Don't show "You subscribed to" bookend on channel creation.
Earlier, on creating a channel we got a "You subscribed to..."
bookend in the message feed after the "channel events" message.

Ideally, the user who created the channel is always considered to
be subscribed before the automated message is sent. So, the bookend
shouldn't be displayed.

This commit fixes the incorrect behaviour.